+++ src/layers/mod.rs
@@ -2,7 +2,7 @@ use std::path::Path;
 
 use xml::attribute::OwnedAttribute;
 
-use crate::{error::TiledError, properties::Properties, util::*, Map, MapTilesetGid, MapWrapper};
+use crate::{error::TiledError, properties::Properties, util::*, Map, MapTilesetGid, MapWrapper, Color};
 
 mod image;
 pub use image::*;
@@ -36,6 +36,7 @@ pub struct LayerData {
     pub parallax_x: f32,
     pub parallax_y: f32,
     pub opacity: f32,
+    pub tint_color: Option<Color>,
     pub properties: Properties,
     pub layer_type: LayerDataType,
 }
@@ -49,10 +50,14 @@ impl LayerData {
         map_path: &Path,
         tilesets: &[MapTilesetGid],
     ) -> Result<Self, TiledError> {
-        let ((opacity, visible, offset_x, offset_y, parallax_x, parallax_y, name, id), ()) = get_attrs!(
+        let (
+            (opacity, tint_color, visible, offset_x, offset_y, parallax_x, parallax_y, name, id),
+            (),
+        ) = get_attrs!(
             attrs,
             optionals: [
                 ("opacity", opacity, |v:String| v.parse().ok()),
+                ("tintcolor", tint_color, |v:String| v.parse().ok()),
                 ("visible", visible, |v:String| v.parse().ok().map(|x:i32| x == 1)),
                 ("offsetx", offset_x, |v:String| v.parse().ok()),
                 ("offsety", offset_y, |v:String| v.parse().ok()),
@@ -89,6 +94,7 @@ impl LayerData {
             parallax_x: parallax_x.unwrap_or(1.0),
             parallax_y: parallax_y.unwrap_or(1.0),
             opacity: opacity.unwrap_or(1.0),
+            tint_color,
             name: name.unwrap_or_default(),
             id: id.unwrap_or(0),
             properties,